Sort names of imports to make order deterministic
At the moment, whenever a file is written the order of imports is randomised due to them being an unordered set. This means that when jsonschema-gentypes is run via pre-commit the file can potentially change, which pre-commit interprets as a "failure".

This change sorts the set alphabetically before output, which means it should be consistent between runs.